added toastr, alerts, modal, datepicker, some checks and layout changes

+++ src/app/MyComponent/todos/todos.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Todo } from '../../Todo';
-import { Tooltip } from 'node_modules/bootstrap/dist/js/bootstrap.esm.min.js'
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-todos',
@@ -10,50 +10,63 @@ import { Tooltip } from 'node_modules/bootstrap/dist/js/bootstrap.esm.min.js'
 export class TodosComponent implements OnInit {
 
   color: string;
-  ngOnInit() {
-    
-  }
+  date: string;
   localItem: string;
   todos: Todo[];
+
+  ngOnInit() {}
+
   constructor() {
-    this.localItem = localStorage.getItem("todos");
+    this.localItem = localStorage.getItem('todos');
     if (this.localItem == null) {
       this.todos = [];
     } else {
       this.todos = JSON.parse(this.localItem);
     }
-    
   }
 
-  // ngOnInit(): void {}
+  deleteTodo(todo: Todo) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire('Deleted!', 'Your todo has been deleted.', 'success');
 
-  deleteTodo(todo: Todo){
-    const index = this.todos.indexOf(todo);
-    this.todos.splice(index, 1);
-    localStorage.setItem("todos", JSON.stringify(this.todos))
+        const index = this.todos.indexOf(todo);
+        this.todos.splice(index, 1);
+        localStorage.setItem('todos', JSON.stringify(this.todos));
+      }
+    });
   }
 
-  addTodo(todo: Todo){
+  addTodo(todo: Todo) {
     this.todos.push(todo);
-    localStorage.setItem("todos", JSON.stringify(this.todos))
-    console.log("inside add todo")
+    localStorage.setItem('todos', JSON.stringify(this.todos));
   }
 
-  strikeTodo(todo: Todo){
+  strikeTodo(todo: Todo) {
     const index = this.todos.indexOf(todo);
-    this.todos[index].active = !this.todos[index].active 
-    localStorage.setItem("todos", JSON.stringify(this.todos))
+    this.todos[index].active = !this.todos[index].active;
+    localStorage.setItem('todos', JSON.stringify(this.todos));
   }
 
-  changeColor(data){
-    console.log(data.todo)
+  changeColor(data) {
+    const index = this.todos.indexOf(data.todo);
+    const col = data.color;
+    this.todos[index].color = col;
+    localStorage.setItem('todos', JSON.stringify(this.todos));
+  }
 
-    // console.log(Object.values(data.todo))
-    const index = this.todos.indexOf(data.todo)
-    const col = data.color
-    this.todos[index].color = col
-    localStorage.setItem("todos", JSON.stringify(this.todos))
+  applyDate(data) {
+    const index = this.todos.indexOf(data.todo);
+    const col = data.date;
+    this.todos[index].date = new Date(col).toISOString().split('T')[0];
+    localStorage.setItem('todos', JSON.stringify(this.todos));
   }
 }
-
-
